+++ src/test/java/com/github/mutare/adventcalendar2019/day5/FileProgramInputSource.java
@@ -10,7 +10,7 @@ import java.util.LinkedList;
 import java.util.stream.Collectors;
 
 public class FileProgramInputSource {
-    public int[] getProgram() throws IOException {
+    public long[] getProgram() throws IOException {
         Collection<Integer> program = new LinkedList<>();
         try (InputStream is = FileProgramInputSource.class.getResourceAsStream("/day5/data")) {
             InputStreamReader inputStreamReader = new InputStreamReader(is);
@@ -20,6 +20,6 @@ public class FileProgramInputSource {
                 program.addAll(Arrays.asList(line.split(",")).stream().map(Integer::parseInt).collect(Collectors.toList()));
             }
         }
-        return program.stream().mapToInt(value -> value.intValue()).toArray();
+        return program.stream().mapToLong(value -> value.longValue()).toArray();
     }
 }
